+++ backend/src/main/java/perf/shop/domain/payment/dao/PaymentRepository.java
@@ -1,9 +1,8 @@
-package perf.shop.domain.payment.repository;
+package perf.shop.domain.payment.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import perf.shop.domain.payment.domain.Payment;
 
 public interface PaymentRepository extends JpaRepository<Payment, Long> {
 
-    boolean existsByOrderId(String orderId);
 }
