+++ backend/src/test/java/perf/shop/domain/payment/application/PaymentServiceTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import perf.shop.domain.payment.repository.PaymentRepository;
+import perf.shop.domain.payment.dao.PaymentRepository;
 
 @ExtendWith(MockitoExtension.class)
 @DisplayName("[단위 테스트] PaymentService")
